Make this explicit

+++ DataPacket.cpp
@@ -56,7 +56,7 @@ unsigned long DataPacket::readRaw(std::uint8_t*   buffer,
 
         // Bump the offset to the next alignment point
         offset_bits = misc::smallestMultipleOfXGreaterOrEqualToY(
-            alignment_bits,
+            static_cast<long>(alignment_bits),
             static_cast<long>(offset_bits));
 
         // This field plus the padding after it
